Removed `reload`, and began semantic versioning and better commit messages.
I deprecated the developer exlusive command `reload`, due to a previous commit modularizing Python, making my approach break. Secondly, this code is also just not recommended in general.

Because I used it so little, I did not expend the effort to create a process manager or other more recommended approaches, but if it becomes something I have need for, I will implement it later.

However for now `reload`, which used to rerun the bot for checking code changes without switching tabs or to reset the state of the bot, has been deprecated.

Currently `kill` should suffice to prevent any erroneous code to run for too long, and can be manually rerun, and also resets the bot's state.